requirements.txt
pdfplumber
python-docx

